--- backend/utils/hashUtils.py
@@ -1,5 +1,0 @@
-import hashlib
-
-
-def hash(text: str):
-    return hashlib.sha256(text.encode("utf-8")).hexdigest()
